refactor(main): extract agent ipc handlers
