+++ src/main/ipcHandlers/agents/index.ts
@@ -1,0 +1,2 @@
+export type { AgentHandlerDeps } from './handlers';
+export { registerAgentHandlers } from './handlers';
